Query Title block: Rely on the editor store to apply the right archive title placeholder
